Added ipynb notebook of colorization by titu1994@
This notebook creates a model that is able to colorize images to a certain extent, combining a deep Convolutional Neural Network architecture. The dataset to be used for the training is MS-COCO. This notebook is imported from https://github.com/titu1994/keras-mobile-colorizer python code. Only minor tweeks to download the data and display images is made.